requirements.txt
bitsandbytes
datasets
accelerate
loralib
transformers
peft
python-dotenv